python/llm-server/requirements.txt
grpcio
grpcio-tools
deepspeed
torch
transformers
huggingface-hub
